solver: separate core functions
